+++ Baker-Biz/Model/ApplePieRecipe.cs
@@ -1,6 +1,6 @@
 using System;
 
-namespace BakerBiz
+namespace BakerBiz.Model
 {
     public class ApplePieRecipe : IPieRecipe
     {
