+++ Baker-Biz/Model/IPieRecipe.cs
@@ -1,6 +1,6 @@
 using System;
 
-namespace BakerBiz
+namespace BakerBiz.Model
 {
     public interface IPieRecipe
     {
