+++ Baker-Biz/Model/IngredientType.cs
@@ -1,0 +1,14 @@
+namespace BakerBiz.Model
+{
+    public enum IngredientType
+    {
+        Apples,
+        Sugar,
+        Flour,
+        Cinnamon,
+        Butter,
+        Blueberries,
+        LemonZest,
+        Milk
+    }
+}
