+++ Baker-Biz/Model/Units.cs
@@ -1,0 +1,12 @@
+namespace BakerBiz.Model
+{
+    public enum Units
+    {
+        number,
+        lbs,
+        tsp,
+        sticks,
+        cup,
+        tbsp
+    }
+}
