+++ Baker-Biz/Program.cs
@@ -1,4 +1,6 @@
 ﻿using System;
+using BakerBiz.Model;
+using BakerBiz.Utilities;
 
 namespace BakerBiz
 {
