--- Baker-Biz/Refactor/IngredientType.cs
@@ -1,7 +1,0 @@
-public enum IngredientType{
-    Apples,
-    Sugar,
-    Flour,
-    Cinnamon,
-    Butter
-}
