--- Baker-Biz/Refactor/Units.cs
@@ -1,6 +1,0 @@
-public enum Units{
-    number,
-    lbs,
-    tsp,
-    sticks
-}
